+++ changelog.md
@@ -10,6 +10,7 @@
 - missing checks in userStatsCache
 - debug commandfiletype using old command names
 - recent activity having the wrong `this.name` value
+- remove unused commands from help 
 
 
 ## [5.0.1] - 2025-05-25