+++ changelog.md
@@ -6,12 +6,15 @@
 
 [commit](https://github.com/sbrstrkkdwmdr/sbrbot)</br>
 
+## [5.0.2] - 2025-05-30
+
+[commit](https://github.com/sbrstrkkdwmdr/sbrbot)</br>
+
 ### Fixed
 - missing checks in userStatsCache
 - debug commandfiletype using old command names
 - recent activity having the wrong `this.name` value
 - remove unused commands from help 
-
 
 ## [5.0.1] - 2025-05-25